Tweaking playlist view to use swipeableitem
